requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.3
